Consolidate ActivityTabModules and ModuleType classes
